+++ internal/provider/framework/provider.go
@@ -5,43 +5,26 @@ package framework
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"iter"
 	"log"
-	"reflect"
 	"slices"
-	"sync"
-	"unique"
 
 	"github.com/hashicorp/terraform-plugin-framework-timetypes/timetypes"
 	"github.com/hashicorp/terraform-plugin-framework-validators/listvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/action"
-	aschema "github.com/hashicorp/terraform-plugin-framework/action/schema"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	datasourceschema "github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/ephemeral"
-	empemeralschema "github.com/hashicorp/terraform-plugin-framework/ephemeral/schema"
 	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/list"
 	"github.com/hashicorp/terraform-plugin-framework/provider"
 	"github.com/hashicorp/terraform-plugin-framework/provider/metaschema"
 	"github.com/hashicorp/terraform-plugin-framework/provider/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
-	resourceschema "github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
-	"github.com/hashicorp/terraform-provider-aws/internal/framework"
 	tffunction "github.com/hashicorp/terraform-provider-aws/internal/function"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
-	inttypes "github.com/hashicorp/terraform-provider-aws/internal/types"
-	tfunique "github.com/hashicorp/terraform-provider-aws/internal/unique"
-	"github.com/hashicorp/terraform-provider-aws/names"
-)
-
-var (
-	resourceSchemasValidated sync.Once
 )
 
 var (
@@ -75,16 +58,6 @@ func NewProvider(ctx context.Context, primary interface{ Meta() any }) (provider
 		primary:            primary,
 		resources:          make([]func() resource.Resource, 0),
 		servicePackages:    primary.Meta().(*conns.AWSClient).ServicePackages(ctx),
-	}
-
-	// Because we try and share resource schemas as much as possible,
-	// we need to ensure that we only validate the resource schemas once.
-	var err error
-	resourceSchemasValidated.Do(func() {
-		err = provider.validateResourceSchemas(ctx)
-	})
-	if err != nil {
-		return nil, err
 	}
 
 	provider.initialize(ctx)
@@ -483,184 +456,4 @@ func (p *frameworkProvider) initialize(ctx context.Context) {
 			}
 		}
 	}
-}
-
-// validateResourceSchemas is called from `New` to validate Terraform Plugin Framework-style resource schemas.
-func (p *frameworkProvider) validateResourceSchemas(ctx context.Context) error {
-	var errs []error
-
-	for sp := range p.servicePackages {
-		for _, dataSourceSpec := range sp.FrameworkDataSources(ctx) {
-			typeName := dataSourceSpec.TypeName
-			inner, err := dataSourceSpec.Factory(ctx)
-
-			if err != nil {
-				errs = append(errs, fmt.Errorf("creating data source type (%s): %w", typeName, err))
-				continue
-			}
-
-			schemaResponse := datasource.SchemaResponse{}
-			inner.Schema(ctx, datasource.SchemaRequest{}, &schemaResponse)
-
-			if err := validateSchemaRegionForDataSource(dataSourceSpec.Region, schemaResponse.Schema); err != nil {
-				errs = append(errs, fmt.Errorf("data source type %q: %w", typeName, err))
-				continue
-			}
-
-			if err := validateSchemaTagsForDataSource(dataSourceSpec.Tags, schemaResponse.Schema); err != nil {
-				errs = append(errs, fmt.Errorf("data source type %q: %w", typeName, err))
-				continue
-			}
-		}
-
-		if v, ok := sp.(conns.ServicePackageWithEphemeralResources); ok {
-			for _, ephemeralResourceSpec := range v.EphemeralResources(ctx) {
-				typeName := ephemeralResourceSpec.TypeName
-				inner, err := ephemeralResourceSpec.Factory(ctx)
-
-				if err != nil {
-					errs = append(errs, fmt.Errorf("creating ephemeral resource type (%s): %w", typeName, err))
-					continue
-				}
-
-				schemaResponse := ephemeral.SchemaResponse{}
-				inner.Schema(ctx, ephemeral.SchemaRequest{}, &schemaResponse)
-
-				if err := validateSchemaRegionForEphemeralResource(ephemeralResourceSpec.Region, schemaResponse.Schema); err != nil {
-					errs = append(errs, fmt.Errorf("ephemeral resource type %q: %w", typeName, err))
-					continue
-				}
-			}
-		}
-
-		if v, ok := sp.(conns.ServicePackageWithActions); ok {
-			for _, actionSpec := range v.Actions(ctx) {
-				typeName := actionSpec.TypeName
-				inner, err := actionSpec.Factory(ctx)
-
-				if err != nil {
-					errs = append(errs, fmt.Errorf("creating action type (%s): %w", typeName, err))
-					continue
-				}
-
-				schemaResponse := action.SchemaResponse{}
-				inner.Schema(ctx, action.SchemaRequest{}, &schemaResponse)
-
-				if err := validateSchemaRegionForAction(actionSpec.Region, schemaResponse.Schema); err != nil {
-					errs = append(errs, fmt.Errorf("action type %q: %w", typeName, err))
-					continue
-				}
-			}
-		}
-
-		for _, resourceSpec := range sp.FrameworkResources(ctx) {
-			typeName := resourceSpec.TypeName
-			inner, err := resourceSpec.Factory(ctx)
-
-			if err != nil {
-				errs = append(errs, fmt.Errorf("creating resource type (%s): %w", typeName, err))
-				continue
-			}
-
-			schemaResponse := resource.SchemaResponse{}
-			inner.Schema(ctx, resource.SchemaRequest{}, &schemaResponse)
-
-			if err := validateSchemaRegionForResource(resourceSpec.Region, schemaResponse.Schema); err != nil {
-				errs = append(errs, fmt.Errorf("resource type %q: %w", typeName, err))
-				continue
-			}
-
-			if err := validateSchemaTagsForResource(resourceSpec.Tags, schemaResponse.Schema); err != nil {
-				errs = append(errs, fmt.Errorf("resource type %q: %w", typeName, err))
-				continue
-			}
-
-			if resourceSpec.Import.WrappedImport {
-				if resourceSpec.Import.SetIDAttr {
-					if _, ok := resourceSpec.Import.ImportID.(inttypes.FrameworkImportIDCreator); !ok {
-						errs = append(errs, fmt.Errorf("resource type %q: importer sets `%s` attribute, but creator isn't configured", resourceSpec.TypeName, names.AttrID))
-						continue
-					}
-				}
-
-				if _, ok := inner.(framework.ImportByIdentityer); !ok {
-					errs = append(errs, fmt.Errorf("resource type %q: cannot configure importer, does not implement %q", resourceSpec.TypeName, reflect.TypeFor[framework.ImportByIdentityer]()))
-					continue
-				}
-			}
-		}
-	}
-
-	return errors.Join(errs...)
-}
-
-func validateSchemaRegionForDataSource(regionSpec unique.Handle[inttypes.ServicePackageResourceRegion], schema datasourceschema.Schema) error {
-	if !tfunique.IsHandleNil(regionSpec) && regionSpec.Value().IsOverrideEnabled {
-		if _, ok := schema.Attributes[names.AttrRegion]; ok {
-			return fmt.Errorf("configured for enhanced regions but defines `%s` attribute in schema", names.AttrRegion)
-		}
-	}
-	return nil
-}
-
-func validateSchemaRegionForEphemeralResource(regionSpec unique.Handle[inttypes.ServicePackageResourceRegion], schema empemeralschema.Schema) error {
-	if !tfunique.IsHandleNil(regionSpec) && regionSpec.Value().IsOverrideEnabled {
-		if _, ok := schema.Attributes[names.AttrRegion]; ok {
-			return fmt.Errorf("configured for enhanced regions but defines `%s` attribute in schema", names.AttrRegion)
-		}
-	}
-	return nil
-}
-
-func validateSchemaRegionForAction(regionSpec unique.Handle[inttypes.ServicePackageResourceRegion], schemaIface any) error {
-	if !tfunique.IsHandleNil(regionSpec) && regionSpec.Value().IsOverrideEnabled {
-		if schema, ok := schemaIface.(aschema.Schema); ok {
-			if _, ok := schema.Attributes[names.AttrRegion]; ok {
-				return fmt.Errorf("configured for enhanced regions but defines `%s` attribute in schema", names.AttrRegion)
-			}
-		}
-	}
-	return nil
-}
-
-func validateSchemaRegionForResource(regionSpec unique.Handle[inttypes.ServicePackageResourceRegion], schema resourceschema.Schema) error {
-	if !tfunique.IsHandleNil(regionSpec) && regionSpec.Value().IsOverrideEnabled {
-		if _, ok := schema.Attributes[names.AttrRegion]; ok {
-			return fmt.Errorf("configured for enhanced regions but defines `%s` attribute in schema", names.AttrRegion)
-		}
-	}
-	return nil
-}
-
-func validateSchemaTagsForDataSource(tagsSpec unique.Handle[inttypes.ServicePackageResourceTags], schema datasourceschema.Schema) error {
-	if !tfunique.IsHandleNil(tagsSpec) {
-		if v, ok := schema.Attributes[names.AttrTags]; ok {
-			if !v.IsComputed() {
-				return fmt.Errorf("`%s` attribute must be Computed", names.AttrTags)
-			}
-		} else {
-			return fmt.Errorf("configured for tags but no `%s` attribute defined in schema", names.AttrTags)
-		}
-	}
-	return nil
-}
-
-func validateSchemaTagsForResource(tagsSpec unique.Handle[inttypes.ServicePackageResourceTags], schema resourceschema.Schema) error {
-	if !tfunique.IsHandleNil(tagsSpec) {
-		if v, ok := schema.Attributes[names.AttrTags]; ok {
-			if v.IsComputed() {
-				return fmt.Errorf("`%s` attribute cannot be Computed", names.AttrTags)
-			}
-		} else {
-			return fmt.Errorf("configured for tags but no `%s` attribute defined in schema", names.AttrTags)
-		}
-		if v, ok := schema.Attributes[names.AttrTagsAll]; ok {
-			if !v.IsComputed() {
-				return fmt.Errorf("`%s` attribute must be Computed", names.AttrTagsAll)
-			}
-		} else {
-			return fmt.Errorf("configured for tags but no `%s` attribute defined in schema", names.AttrTagsAll)
-		}
-	}
-	return nil
 }
