+++ internal/provider/framework/provider_test.go
@@ -4,11 +4,27 @@
 package framework
 
 import (
+	"context"
+	"errors"
+	"fmt"
+	"reflect"
 	"testing"
+	"unique"
 
+	"github.com/hashicorp/terraform-plugin-framework/action"
+	aschema "github.com/hashicorp/terraform-plugin-framework/action/schema"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	datasourceschema "github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/ephemeral"
+	ephemeralschema "github.com/hashicorp/terraform-plugin-framework/ephemeral/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
+	resourceschema "github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-provider-aws/internal/conns"
+	"github.com/hashicorp/terraform-provider-aws/internal/framework"
 	"github.com/hashicorp/terraform-provider-aws/internal/provider/sdkv2"
+	inttypes "github.com/hashicorp/terraform-provider-aws/internal/types"
+	tfunique "github.com/hashicorp/terraform-provider-aws/internal/unique"
+	"github.com/hashicorp/terraform-provider-aws/names"
 )
 
 func TestProviderInit(t *testing.T) {
@@ -54,29 +70,6 @@ func TestProviderInit(t *testing.T) {
 // 	}
 // }
 
-func BenchmarkFrameworkProvider_validateResourceSchemas(b *testing.B) {
-	ctx := b.Context()
-	primary, err := sdkv2.NewProvider(ctx)
-	if err != nil {
-		b.Fatalf("Initializing SDKv2 provider: %s", err)
-	}
-	p, err := NewProvider(ctx, primary)
-	if err != nil {
-		b.Fatalf("Initializing Framework provider: %s", err)
-	}
-
-	provider := p.(*frameworkProvider)
-
-	// Reset memory counters to zero, so that we only measure the schema validation.
-	b.ResetTimer()
-	for b.Loop() {
-		err := provider.validateResourceSchemas(ctx)
-		if err != nil {
-			b.Fatalf("Validating resource schemas: %s", err)
-		}
-	}
-}
-
 func BenchmarkFrameworkProvider_SchemaInitialization_DataSource(b *testing.B) {
 	ctx := b.Context()
 	primary, err := sdkv2.NewProvider(ctx)
@@ -119,4 +112,184 @@ func BenchmarkFrameworkProvider_SchemaInitialization_Resource(b *testing.B) {
 			r.Schema(ctx, resource.SchemaRequest{}, &resource.SchemaResponse{})
 		}
 	}
+}
+
+// validateResourceSchemas is called in a unit test to validate Terraform Plugin Framework-style resource schemas.
+func (p *frameworkProvider) validateResourceSchemas(ctx context.Context) error {
+	var errs []error
+
+	for sp := range p.servicePackages {
+		for _, dataSourceSpec := range sp.FrameworkDataSources(ctx) {
+			typeName := dataSourceSpec.TypeName
+			inner, err := dataSourceSpec.Factory(ctx)
+
+			if err != nil {
+				errs = append(errs, fmt.Errorf("creating data source type (%s): %w", typeName, err))
+				continue
+			}
+
+			schemaResponse := datasource.SchemaResponse{}
+			inner.Schema(ctx, datasource.SchemaRequest{}, &schemaResponse)
+
+			if err := validateSchemaRegionForDataSource(dataSourceSpec.Region, schemaResponse.Schema); err != nil {
+				errs = append(errs, fmt.Errorf("data source type %q: %w", typeName, err))
+				continue
+			}
+
+			if err := validateSchemaTagsForDataSource(dataSourceSpec.Tags, schemaResponse.Schema); err != nil {
+				errs = append(errs, fmt.Errorf("data source type %q: %w", typeName, err))
+				continue
+			}
+		}
+
+		if v, ok := sp.(conns.ServicePackageWithEphemeralResources); ok {
+			for _, ephemeralResourceSpec := range v.EphemeralResources(ctx) {
+				typeName := ephemeralResourceSpec.TypeName
+				inner, err := ephemeralResourceSpec.Factory(ctx)
+
+				if err != nil {
+					errs = append(errs, fmt.Errorf("creating ephemeral resource type (%s): %w", typeName, err))
+					continue
+				}
+
+				schemaResponse := ephemeral.SchemaResponse{}
+				inner.Schema(ctx, ephemeral.SchemaRequest{}, &schemaResponse)
+
+				if err := validateSchemaRegionForEphemeralResource(ephemeralResourceSpec.Region, schemaResponse.Schema); err != nil {
+					errs = append(errs, fmt.Errorf("ephemeral resource type %q: %w", typeName, err))
+					continue
+				}
+			}
+		}
+
+		if v, ok := sp.(conns.ServicePackageWithActions); ok {
+			for _, actionSpec := range v.Actions(ctx) {
+				typeName := actionSpec.TypeName
+				inner, err := actionSpec.Factory(ctx)
+
+				if err != nil {
+					errs = append(errs, fmt.Errorf("creating action type (%s): %w", typeName, err))
+					continue
+				}
+
+				schemaResponse := action.SchemaResponse{}
+				inner.Schema(ctx, action.SchemaRequest{}, &schemaResponse)
+
+				if err := validateSchemaRegionForAction(actionSpec.Region, schemaResponse.Schema); err != nil {
+					errs = append(errs, fmt.Errorf("action type %q: %w", typeName, err))
+					continue
+				}
+			}
+		}
+
+		for _, resourceSpec := range sp.FrameworkResources(ctx) {
+			typeName := resourceSpec.TypeName
+			inner, err := resourceSpec.Factory(ctx)
+
+			if err != nil {
+				errs = append(errs, fmt.Errorf("creating resource type (%s): %w", typeName, err))
+				continue
+			}
+
+			schemaResponse := resource.SchemaResponse{}
+			inner.Schema(ctx, resource.SchemaRequest{}, &schemaResponse)
+
+			if err := validateSchemaRegionForResource(resourceSpec.Region, schemaResponse.Schema); err != nil {
+				errs = append(errs, fmt.Errorf("resource type %q: %w", typeName, err))
+				continue
+			}
+
+			if err := validateSchemaTagsForResource(resourceSpec.Tags, schemaResponse.Schema); err != nil {
+				errs = append(errs, fmt.Errorf("resource type %q: %w", typeName, err))
+				continue
+			}
+
+			if resourceSpec.Import.WrappedImport {
+				if resourceSpec.Import.SetIDAttr {
+					if _, ok := resourceSpec.Import.ImportID.(inttypes.FrameworkImportIDCreator); !ok {
+						errs = append(errs, fmt.Errorf("resource type %q: importer sets `%s` attribute, but creator isn't configured", resourceSpec.TypeName, names.AttrID))
+						continue
+					}
+				}
+
+				if _, ok := inner.(framework.ImportByIdentityer); !ok {
+					errs = append(errs, fmt.Errorf("resource type %q: cannot configure importer, does not implement %q", resourceSpec.TypeName, reflect.TypeFor[framework.ImportByIdentityer]()))
+					continue
+				}
+			}
+		}
+	}
+
+	return errors.Join(errs...)
+}
+
+func validateSchemaRegionForDataSource(regionSpec unique.Handle[inttypes.ServicePackageResourceRegion], schema datasourceschema.Schema) error {
+	if !tfunique.IsHandleNil(regionSpec) && regionSpec.Value().IsOverrideEnabled {
+		if _, ok := schema.Attributes[names.AttrRegion]; ok {
+			return fmt.Errorf("configured for enhanced regions but defines `%s` attribute in schema", names.AttrRegion)
+		}
+	}
+	return nil
+}
+
+func validateSchemaRegionForEphemeralResource(regionSpec unique.Handle[inttypes.ServicePackageResourceRegion], schema ephemeralschema.Schema) error {
+	if !tfunique.IsHandleNil(regionSpec) && regionSpec.Value().IsOverrideEnabled {
+		if _, ok := schema.Attributes[names.AttrRegion]; ok {
+			return fmt.Errorf("configured for enhanced regions but defines `%s` attribute in schema", names.AttrRegion)
+		}
+	}
+	return nil
+}
+
+func validateSchemaRegionForAction(regionSpec unique.Handle[inttypes.ServicePackageResourceRegion], schemaIface any) error {
+	if !tfunique.IsHandleNil(regionSpec) && regionSpec.Value().IsOverrideEnabled {
+		if schema, ok := schemaIface.(aschema.Schema); ok {
+			if _, ok := schema.Attributes[names.AttrRegion]; ok {
+				return fmt.Errorf("configured for enhanced regions but defines `%s` attribute in schema", names.AttrRegion)
+			}
+		}
+	}
+	return nil
+}
+
+func validateSchemaRegionForResource(regionSpec unique.Handle[inttypes.ServicePackageResourceRegion], schema resourceschema.Schema) error {
+	if !tfunique.IsHandleNil(regionSpec) && regionSpec.Value().IsOverrideEnabled {
+		if _, ok := schema.Attributes[names.AttrRegion]; ok {
+			return fmt.Errorf("configured for enhanced regions but defines `%s` attribute in schema", names.AttrRegion)
+		}
+	}
+	return nil
+}
+
+func validateSchemaTagsForDataSource(tagsSpec unique.Handle[inttypes.ServicePackageResourceTags], schema datasourceschema.Schema) error {
+	if !tfunique.IsHandleNil(tagsSpec) {
+		if v, ok := schema.Attributes[names.AttrTags]; ok {
+			if !v.IsComputed() {
+				return fmt.Errorf("`%s` attribute must be Computed", names.AttrTags)
+			}
+		} else {
+			return fmt.Errorf("configured for tags but no `%s` attribute defined in schema", names.AttrTags)
+		}
+	}
+	return nil
+}
+
+func validateSchemaTagsForResource(tagsSpec unique.Handle[inttypes.ServicePackageResourceTags], schema resourceschema.Schema) error {
+	if !tfunique.IsHandleNil(tagsSpec) {
+		if v, ok := schema.Attributes[names.AttrTags]; ok {
+			if v.IsComputed() {
+				return fmt.Errorf("`%s` attribute cannot be Computed", names.AttrTags)
+			}
+		} else {
+			return fmt.Errorf("configured for tags but no `%s` attribute defined in schema", names.AttrTags)
+		}
+		if v, ok := schema.Attributes[names.AttrTagsAll]; ok {
+			if !v.IsComputed() {
+				return fmt.Errorf("`%s` attribute must be Computed", names.AttrTagsAll)
+			}
+		} else {
+			return fmt.Errorf("configured for tags but no `%s` attribute defined in schema", names.AttrTagsAll)
+		}
+	}
+	return nil
 }
